backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.0
decorator==4.4.2
dj-database-url==0.5.0
drf-yasg==1.20.0
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
future==0.18.2
geocoder==1.38.1
gunicorn==20.0.4
idna==2.10
Pillow==7.2.0
psycopg2-binary==2.8.5
pycparser==2.20
pycryptodome==3.9.8
python-dateutil==2.8.1
pytz==2020.1
ratelim==0.1.6
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
